Add logging mechanism
This commit aims to generate a logging system. This is useful for the
massive amounts of logging required to sort out issues rising with the
usage of extremely large databases on prod. In the future only important
messages, like status updates on certain actions will get printed into
the console, while replica of the same algorithms like sorting certain
students will get placed into the INFO category of the log.

The log is titled "sdb-log.txt" and will get created at the start of the
script!