Add basic mode input and rename sibank_converter
Renamed sibank_converter.py into sibank_classlist_converter.py for better
understanding of the different naming schemes. In main.py added a basic
check for the mode, this can be improved later on, as the code isnt
optimized. It currently impacts nothing, need to implement functions for
the converters first.