Add functionality for classlist --> Iserv userlist
In main.py was the function from sibank_classlist_converter.py added
for better clearity in the codebase. In install.py  a check was implemented
to verify that the required files for the conversion are accessable.
The pandas framework seems not implemented into the base Python packages,
so users will need to pip install it first.