Add filtering for empty and wrong IDENTNUMMER in mailinglist
This commit aims at a problem rising when using the Iserv-db as it has
some dummy ID-Numbers which are generated by Iserv. In the mailinglist-
function I added a check that looks if "IDENTNUMMER" is numeric. If it isn't,
it doesn't get added to the dataframe. Same logic should apply when no
ID-Number is available. (Last one wasn't tested by me)