Change program behavior on execution
This commit will change the whole program into automatic use. As I already
implemented checks if certain files are in place, the program should be
pretty failsafe so it should be safe to fully rely on those checks in the
install.py file. Now the programm will only start the filechecks and the
functions depending on them. After that the program will automatically
change the mode to the next one.
Also updated the messages that were displayed by install.py --> they Now
show the files that are required for the next operations